Folder restructure, added controller and services folder

+++ src/main/java/com/spring/Healthcare/repository/AdminRepository.java
@@ -1,10 +1,10 @@
 package com.spring.Healthcare.repository;
 
-import com.spring.Healthcare.model.Admin;
+import com.spring.Healthcare.data.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
-    // Custom queries can be added here if needed
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
+
 }
